requirements.txt
fastmcp>=2.8.1
python-docx>=1.1.2